2013-05-11: fixed rpm.spec rpm.sh

+++ rpmbuild/SPECS/coreutils.spec
@@ -25,7 +25,7 @@ FORCE_UNSAFE_CONFIGURE=1  ./configure \
 	--enable-no-install-program=kill,uptime
 make %{?_smp_mflags}
 %install
-rm -rf %{buildroot}
+[ %{buildroot} != "/"] && rm -rf %{buildroot}/*
 make DESTDIR=%{buildroot} install
 install -vdm 755 %{buildroot}/bin
 install -vdm 755 %{buildroot}%{_sbindir}
